LayoutTests/imported/w3c/web-platform-tests/screen-orientation/lock-bad-argument-expected.txt


PASS screen.orientation.lock() must throw given invalid input.
PASS screen.orientation.lock() must throw when the input is missing.
